% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Add together two numbers.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{gra}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}

