% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
the greatest common divisor of \code{x} and \code{y}.
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\description{
Find the greatest common divisor of two numbers.
}
\examples{
euclidian(123612, 13892347912)
euclidian(100, 1000)
}

