\name{wiki_graph}
\alias{wiki_graph}
\docType{data}
\title{
wiki_graph
}
\description{
test data
}
\usage{data(wiki_graph)}
\format{
  A data frame with 18 observations on the following 3 variables.
  \describe{
    \item{\code{v1}}{a numeric vector}
    \item{\code{v2}}{a numeric vector}
    \item{\code{w}}{a numeric vector}
  }
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
