\name{My first package}
\alias{myfirstpackage}
\alias{myfirstpackage}
\docType{package}
\title{
{My First Package}
}
\description{
{My first package to lab 3}
}
\details{

Euclidean function\cr
\cr
Disjkstra function \cr
\cr
Wiki_gragh dataframe
}
\author{
{Yixuan}\cr\cr
{Violeta}

Maintainers: {Yixuan and Violeta}
}
\references{
}
\keyword{ package }

\examples{
euclidean(24, 6)
dijkstra(wiki_graph, 1)
}
