% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm is an algorithm for finding the shortest
paths between nodes in a graph.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data with v1,v2 & w.}

\item{init_node}{A number.}
}
\value{
the greatest common divisor of \code{x} and \code{y}.
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\description{
Dijkstra's algorithm is an algorithm for finding the shortest
paths between nodes in a graph.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}

