\name{My first package}
\alias{myfirstpackage}
\alias{myfirstpackage}
\docType{package}
\title{
{My First Package}
}
\description{
{My first package to lab 3}
}
<<<<<<< HEAD
=======
\details{

Euclidean function\cr
\cr
Disjkstra function \cr
\cr
Wiki_gragh dataframe
}
>>>>>>> 772c2c6fcc73f9c8327aee7a67882ad09fb8c607
\author{
{Yixuan}\cr\cr
{Violeta}

Maintainers: {Yixuan and Violeta}
}
<<<<<<< HEAD
\keyword{ package }

=======
\references{
}
\keyword{ package }

\examples{
euclidean(24, 6)
dijkstra(wiki_graph, 1)
}
>>>>>>> 772c2c6fcc73f9c8327aee7a67882ad09fb8c607
