\name{yixuan-package}
\alias{yixuan}
\docType{package}
\title{
\packageTitle{yixuan}
}
\description{
\packageDescription{yixuan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{yixuan}
\packageIndices{yixuan}
Two functions include for testing how to build a package
}
\author{
\packageAuthor{Yixuan Xu & Violeta García Sáez}

}

\keyword{ euclidean, dijkstra }

