\name{yixuan-package}
\alias{yixuan-package}
\alias{yixuan}
\docType{package}
\title{
\packageTitle{yixuan}
}
\description{
\packageDescription{yixuan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{yixuan}
\packageIndices{yixuan}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{yixuan}

Maintainer: \packageMaintainer{yixuan}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
